test-updates updating tests from react 19 updates
