+++ src/components/ParaText/ParaText.test.js
@@ -1,34 +1,10 @@
-import { createRoot } from "react-dom/client";
-import React from "react";
-import { render, unmountComponentAtNode } from "react-dom";
-import { act } from "react";
+import React from 'react';
+import { act, screen, render } from '@testing-library/react';
 import ParaText from './ParaText';
 
-let container = null;
-beforeEach(() => {
-  // setup a DOM element as a render target
-  container = document.createElement("div");
-  document.body.appendChild(container);
-});
-
-afterEach(() => {
-  const root = createRoot(container);
-  root.unmount();
-  container.remove();
-  container = null;
-});
-
-it('renders when passing a string and when passing an empty string', () => {
-  act(() => {
-    const root = createRoot(container);
-    root.render(<ParaText>Hello World!</ParaText>);
+describe('ParaText component', () => {
+  it('renders when passing a string and when passing an empty string', () => {
+    render(<ParaText>Hello World!</ParaText>);
+    expect(screen.getByText('Hello World!')).toBeInTheDocument();
   });
-  expect(container.textContent).toBe("Hello World!");
-
-  act(() => {
-    const root = createRoot(container);
-    root.render(<ParaText></ParaText>);
-  });
-  expect(container.textContent).toBe("");
-  
 });
